src/requirements.txt
amqp==2.5.0
anyjson==0.3.3
billiard==3.6.0.0
wheel==0.36.2
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-timezone-field==3.0
djangorestframework==3.9.4
idna==2.8
kombu==4.6.3
numpy==1.16.4
opencv-python==4.5.1.48
Pillow==6.0.0
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
virtualenv==16.6.0
